practice/11.deploy-and-template/kubelint/requirements.txt
pathspec==0.5.9
PyYAML==5.1
supermutes==0.2.5
yamllint==1.15.0
